# app/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "eth-tools API",
    "description": "Self-maintaining runtime for ERC-8004 trustless agents.",
    "license": {
      "name": "MIT"
    },
    "version": "0.0.1"
  },
  "servers": [
    {
      "url": "https://eth-tools.dev",
      "description": "Production"
    },
    {
      "url": "https://preview.eth-tools.dev",
      "description": "Preview"
    },
    {
      "url": "http://localhost:3000",
      "description": "Local dev"
    }
  ],
  "paths": {
    "/api/v1/access/check": {
      "post": {
        "tags": [
          "access"
        ],
        "operationId": "access_check",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AccessRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Boolean allow/deny + one-line reason",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccessCheckResponse"
                }
              }
            }
          },
          "400": {
            "description": "bad input (address / agent_id / chain)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          },
          "404": {
            "description": "chain not indexed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          },
          "429": {
            "description": "rate limited",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          },
          "500": {
            "description": "internal error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/access/explain": {
      "post": {
        "tags": [
          "access"
        ],
        "operationId": "access_explain",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AccessRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Verbose policy breakdown",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccessExplainResponse"
                }
              }
            }
          },
          "400": {
            "description": "bad input",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          },
          "404": {
            "description": "chain not indexed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          },
          "429": {
            "description": "rate limited",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          },
          "500": {
            "description": "internal error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AccessCheckResponse": {
        "type": "object",
        "required": [
          "allowed",
          "reason"
        ],
        "properties": {
          "allowed": {
            "type": "boolean"
          },
          "reason": {
            "type": "string",
            "description": "One-line summary; details under `/access/explain`."
          }
        }
      },
      "AccessExplainResponse": {
        "type": "object",
        "required": [
          "allowed",
          "reason",
          "steps"
        ],
        "properties": {
          "allowed": {
            "type": "boolean"
          },
          "reason": {
            "type": "string"
          },
          "steps": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AccessStep"
            },
            "description": "Ordered breakdown of which checks ran, what each returned, and the\nrule that produced the final answer."
          }
        }
      },
      "AccessRequest": {
        "type": "object",
        "required": [
          "chain",
          "agent_id",
          "requester_address",
          "action"
        ],
        "properties": {
          "action": {
            "type": "string",
            "description": "Free-form action name (`invoke`, `read_manifest`, …). The policy\nengine is the source of truth — handler echoes back."
          },
          "agent_id": {
            "type": "string",
            "description": "uint256 as decimal string."
          },
          "chain": {
            "type": "string"
          },
          "requester_address": {
            "type": "string",
            "description": "0x-prefixed 20-byte hex."
          }
        }
      },
      "AccessStep": {
        "type": "object",
        "required": [
          "check",
          "ok",
          "detail"
        ],
        "properties": {
          "check": {
            "type": "string"
          },
          "detail": {
            "type": "string"
          },
          "ok": {
            "type": "boolean"
          }
        }
      },
...
      "ApiErrorBody": {
        "type": "object",
        "description": "Top-level error envelope returned by every non-2xx response.",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ApiErrorPayload"
          }
        }
      },
      "ApiErrorPayload": {
        "type": "object",
        "required": [
          "code",
          "policy_version",
          "evaluator"
        ],
        "properties": {
          "code": {
            "type": "string",
            "example": "AGENT_NOT_FOUND"
          },
          "evaluator": {
            "type": "string",
            "example": "api.lookup"
          },
          "override_hint": {
            "type": [
              "string",
              "null"
            ]
          },
          "policy_version": {
            "type": "string",
            "example": "v1"
          }
        }
      },
...
    }
  },
...
}